👌 Move migration code to own pkg
