+++ upgrades/0.0.97.seperate-ns-from-app/pkg/migration/migrations.go
@@ -1,4 +1,4 @@
-package upgrade
+package migration
 
 import (
 	"fmt"
@@ -9,22 +9,6 @@ import (
 	"github.com/oslokommune/okctl-upgrade/upgrades/0.0.97.seperate-ns-from-app/pkg/lib/manifest/apis/okctl.io/v1alpha1"
 	"github.com/spf13/afero"
 )
-
-func migrateExistingNamespacesToDir(fs *afero.Afero, cluster v1alpha1.Cluster, absoluteRepositoryRoot string) error {
-	apps, err := getApplicationsInCluster(fs, cluster, absoluteRepositoryRoot)
-	if err != nil {
-		return fmt.Errorf("scanning for : %w", err)
-	}
-
-	for _, app := range apps {
-		err = migrateApplication(fs, cluster, absoluteRepositoryRoot, app)
-		if err != nil {
-			return fmt.Errorf("migrating %s: %w", app, err)
-		}
-	}
-
-	return nil
-}
 
 func migrateApplication(fs *afero.Afero, cluster v1alpha1.Cluster, absoluteRepositoryRoot string, appName string) error {
 	absoluteNamespacesDir := path.Join(absoluteRepositoryRoot, paths.RelativeNamespacesDir(cluster))
