+++ upgrades/0.0.97.seperate-ns-from-app/pkg/migration/migrations_helpers.go
@@ -1,4 +1,4 @@
-package upgrade
+package migration
 
 import (
 	"fmt"
